// src/components/screens/LancamentoInicial/Relatorios/RelatorioAlimentacoesServidas/types.ts
export type IFiltros = {
  mes?: string;
  dres?: Array<string>;
  lotes?: Array<string>;
  subprefeituras?: Array<string>;
  tipos_unidades?: Array<string>;
  unidades_educacionais?: Array<string>;
  tipos_alimentacao?: Array<string>;
  faixas_etarias?: Array<string>;
  periodo_lancamento_de?: string;
  periodo_lancamento_ate?: string;
};